requirements.txt
paddlenlp>=2.6.0
paddlepaddle>=2.5.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
jsonlines>=3.0.0
click>=8.0.0
loguru>=0.6.0
tqdm>=4.62.0
onnx>=1.12.0
onnxruntime>=1.12.0
optuna>=3.0.0
huggingface-hub>=0.14.0
colorama>=0.4.6
